# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnHandListResponse": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/OnHandRowResponse"
            },
            "title": "Rows",
            "type": "array"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/OnHandSummaryResponse"
            },
            "title": "Summaries",
            "type": "array"
          }
        },
        "title": "OnHandListResponse",
        "type": "object"
      },
      "OnHandRowResponse": {
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_kind": {
            "enum": [
              "material",
              "supply",
              "product"
            ],
            "title": "Entity Kind",
            "type": "string"
          },
          "location_id": {
            "format": "uuid",
            "title": "Location Id",
            "type": "string"
          },
          "on_hand": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "On Hand",
            "type": "string"
          }
        },
        "required": [
          "entity_kind",
          "entity_id",
          "location_id",
          "on_hand"
        ],
        "title": "OnHandRowResponse",
        "type": "object"
      },
      "OnHandSummaryResponse": {
        "description": "Aggregated across locations for a single entity.",
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_kind": {
            "enum": [
              "material",
              "supply",
              "product"
            ],
            "title": "Entity Kind",
            "type": "string"
          },
          "per_location": {
            "additionalProperties": {
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
              "type": "string"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "title": "Per Location",
            "type": "object"
          },
          "total_on_hand": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total On Hand",
            "type": "string"
          }
        },
        "required": [
          "entity_kind",
          "entity_id",
          "total_on_hand"
        ],
        "title": "OnHandSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "voxel-ledger-backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inventory/on-hand": {
      "get": {
        "description": "Return ``inventory_on_hand`` rows, optionally filtered.\n\nDefault shape: a list of per-(entity, location) rows AND a list of\nper-entity summaries (total + per-location map). The caller can use\neither depending on whether they need the breakdown or the rollup.",
        "operationId": "list_on_hand_api_v1_inventory_on_hand_get",
        "parameters": [
          {
            "in": "query",
            "name": "entity_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "material",
                    "supply",
                    "product"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Kind"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnHandListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List On Hand",
        "tags": [
          "inventory-on-hand"
        ]
      }
    },
...
  }
}